update initialization of mfpLayer
