+++ src/modules/olMap/handler/geolocation/styleUtils.js
@@ -1,7 +1,6 @@
 import { Style, Stroke, Fill, Circle as CircleStyle } from 'ol/style';
 import { easeOut } from 'ol/easing';
 import { getVectorContext } from 'ol/render';
-import { DEVICE_PIXEL_RATIO } from 'ol/has';
 
 export const geolocationStyleFunction = () => [new Style({
 	fill: new Fill({
@@ -20,16 +19,6 @@ export const geolocationStyleFunction = () => [new Style({
 			color: [255, 255, 255, 1],
 			width: 2
 		})
-	})
-})];
-
-export const mfpBoundaryStyleFunction = () => [new Style({
-	fill: new Fill({
-		color: [255, 0, 0, 0.1]
-	}),
-	stroke: new Stroke({
-		color: [255, 0, 0, 0.9],
-		width: 3
 	})
 })];
 
@@ -81,47 +70,3 @@ export const createAnimateFunction = (map, feature, endCallback) => {
 	return animate;
 };
 
-export const createMapMaskFunction = (map, feature) => {
-	const renderMask = (event) => {
-		const vectorContext = getVectorContext(event);
-
-		const geometryToMask = feature.getGeometry().clone();
-
-		const size = this._map.getSize();
-		const width = size[0] * DEVICE_PIXEL_RATIO;
-		const height = size[1] * DEVICE_PIXEL_RATIO;
-
-
-		vectorContext.beginPath();
-		// the outside polygon -> clockwise
-		vectorContext.moveTo(0, 0);
-		vectorContext.lineTo(width, 0);
-		vectorContext.lineTo(width, height);
-		vectorContext.lineTo(0, height);
-		vectorContext.lineTo(0, 0);
-		vectorContext.closePath();
-
-		// the hole (inner polygon) -> counter-clockwise
-		const hole = geometryToMask.getCoordinates(true);
-		hole.forEach((element, index, array) => {
-			if (index === 0) { // first
-				vectorContext.moveTo(element[0], element[1]);
-			}
-			if (index === array.length - 1) { //last
-				const firstElement = array[0];
-				vectorContext.lineTo(element[0], element[1]);
-				vectorContext.lineTo(firstElement[0], firstElement[1]);
-			}
-			vectorContext.lineTo(element[0], element[1]);
-		});
-		vectorContext.closePath();
-
-		vectorContext.fillStyle = 'rgba(0,5,25,0.75)';
-		vectorContext.fill();
-
-		// tell OpenLayers to continue postrender animation
-		map.render();
-	};
-	return renderMask;
-
-};
